[Feat] ChipMenuData의 categoryList, regionData 생성 및 타입 지정
